+++ app/+mag/+app/+control/Field.m
@@ -18,7 +18,7 @@ classdef Field < mag.app.control.Control & mag.app.mixin.StartEndDate
             this.Layout = this.createDefaultGridLayout();
 
             % Start and end dates.
-            this.addStartEndDateButtons(this.Layout, 1, 2);
+            this.addStartEndDateButtons(this.Layout, StartDateRow = 1, EndDateRow = 2);
 
             % Events.
             eventsLabel = uilabel(this.Layout, Text = "Events:");